requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose==3.3.0

pytest==8.3.2
pytest-cov==5.0.0
httpx==0.27.0

mypy==1.11.2
pylint==3.2.7

email-validator==2.2.0
python-multipart==0.0.9
